Move shared components to components/shared
